Ensure game object is saved after checking for existing matching games in poll_api command

+++ GHLWebsiteDjango/GHLWebsiteApp/management/commands/poll_api.py
@@ -151,9 +151,9 @@ class Command(BaseCommand):
                     h_team_gf=int(h_team_gf),
                     season_num=season_instance,
                 )
-                game_obj.save()
-                if matching_games.exists():
+                if matching_games.exists() and season_instance.season_num == get_seasonSetting():
                     matching_games.first().delete()
+                game_obj.save()
                 self.stdout.write(f"Created new game {game_num}")
                 
                 # Parse team stats
